Add text to comment item

+++ plugins/robots/common/twoDModel/src/engine/view/scene/twoDModelScene.cpp
@@ -657,7 +657,8 @@ void TwoDModelScene::deleteWithCommand(const QStringList &worldItems
 
 void TwoDModelScene::keyPressEvent(QKeyEvent *event)
 {
-	if (event->matches(QKeySequence::Delete) || event->key() == Qt::Key_Backspace) {
+	if ((event->matches(QKeySequence::Delete) || event->key() == Qt::Key_Backspace)
+			&& !selectedItems().isEmpty()) {
 		deleteSelectedItems();
 	} else if (event->matches(QKeySequence::Copy)) {
 		copySelectedItems();
